Split config client network parsing
